final
this project is closed.
the connectivity is based on a Delaunay traingulation/Voronoi diagram.
The graphs produced by these methods are not guaranteed to contain the
optimal cycle (i.e., counterexamples exist s.t. the optimal cycle
contains an edge that is not in the Delaunay traingulation).

The readme indicates problems that may be encountered on reopening the
project:
1.) make sure ./glut is listed in the project include folders
2.) The code/solution expects VS 2015+. An earlier version will likely
throw an error: "cannot specify explicit initializer for arrays".
Workaround is possible by modifying the code, but has not been
implemented as this project is retired.